Refactor; extract area tags.
